Scraper base updated and types moved to own files
Minor structure change for the scraper base as "Init()" is now considered
a Parser step and thus handled by "Parse()".  "BuildQuery()" prepares
CURL queries using partially filled in information.

The core types for the scrapers has been moved to it's own file to make it
easier to read the scraper base class definition.